PROTOCOL_DARKPLACES5
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@4499 d7cf8633-e32d-0410-b094-e92efae38249

+++ cgamevm.c
@@ -87,7 +87,7 @@ unsigned char CGVM_MSG_ReadByte(void)
 
 short CGVM_MSG_ReadShort(void)
 {
-	int num;
+	short num;
 	num = CGVM_MSG_ReadByte() | (CGVM_MSG_ReadByte() << 8);
 	return num;
 }
